requirements.txt
flask
pillow
imagehash
sqlalchemy
python-magic